reporting/streamlit/requirements.txt
streamlit 
matplotlib
networkx
google-cloud-secret-manager
numpy
pandas
duckdb
